@@ -1,57 +1,57 @@
 {
-    "name": "sorteador-credilab-ng",
-    "version": "1.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "node server.js",
-        "build": "ng build --configuration production",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e"
-    },
-    "engines": {
-        "node": "12.18.2",
-        "npm": "6.14.5"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^13.4.0",
-        "@angular/common": "^13.4.0",
-        "@angular/compiler": "^13.4.0",
-        "@angular/core": "^13.4.0",
-        "@angular/forms": "^13.4.0",
-        "@angular/platform-browser": "^13.4.0",
-        "@angular/platform-browser-dynamic": "^13.4.0",
-        "@angular/router": "^13.4.0",
-        "bootstrap": "^4.5.3",
-        "express": "^4.17.1",
-        "jquery": "^3.5.1",
-        "lodash": "^4.17.21",
-        "moment": "^2.29.1",
-        "ngx-bootstrap": "^6.2.0",
-        "ngx-toastr": "^13.2.0",
-        "rxjs": "~6.6.0",
-        "tslib": "^2.0.0",
-        "zone.js": "~0.11.4"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^13.3.11",
-        "@angular/cli": "^13.3.11",
-        "@angular/compiler-cli": "^13.4.0",
-        "@types/jasmine": "~3.6.0",
-        "@types/lodash": "^4.14.170",
-        "@types/node": "^12.11.1",
-        "codelyzer": "^6.0.0",
-        "jasmine-core": "~3.6.0",
-        "jasmine-spec-reporter": "~5.0.0",
-        "karma": "~6.4.4",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.0.3",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "^1.5.0",
-        "protractor": "~7.0.0",
-        "ts-node": "~8.3.0",
-        "tslint": "~6.1.0",
-        "typescript": "^4.6.4"
-    }
+  "name": "sorteador-credilab-ng",
+  "version": "1.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "node server.js",
+    "build": "ng build --configuration production",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "engines": {
+    "node": "12.18.2",
+    "npm": "6.14.5"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "bootstrap": "^4.5.3",
+    "express": "^4.17.1",
+    "jquery": "^3.5.1",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-toastr": "^13.2.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@types/jasmine": "~3.6.0",
+    "@types/lodash": "^4.14.170",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.4",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "protractor": "~7.0.0",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "^4.6.4"
+  }
 }
